+++ core/src/routes/(modules)/options/text.svelte
@@ -6,6 +6,9 @@
 	import Optionsbutton from "../../uicomponents/optionsbutton.svelte";
 	import Textarea from "../../uicomponents/textarea.svelte";
 	import Textinput from "../../uicomponents/textinput.svelte";
+	import Select from "../../uicomponents/select.svelte";
+
+    
 
 
     let globalComponentCollection = $globalComponentCollectionStore;
@@ -27,6 +30,7 @@
      * @property {number} [tabindex] - Specifies the tab order of the element
      * @property {string} [title] - Specifies extra information about the element (displayed as a tooltip)
      * @property {string} [text] - Value of Text Element
+     * @property {string} [htmltag] - HTML Tag of Text Element
      */
 
     /**
@@ -45,6 +49,21 @@
      */
     let activeElement;
 
+    // /**
+    //  * Global class collection in this file.
+    //  * @type string
+    //  */
+    // let classGlobal;
+
+    // $: classGlobal, (()=>{
+    //     if(activeElement && loaded){
+    //         if(activeElement.data){
+    //             activeElement.data.class = classGlobal;
+    //             if(classInput != classGlobal) classInput = classGlobal;
+    //             updateEditor();
+    //         }
+    //     }
+    // })();
 
     /**
      * Loads element data from JSON 
@@ -54,6 +73,27 @@
 
         textInput = activeElement?.data?.text != undefined ? activeElement?.data?.text : "";
         classInput = activeElement?.data?.class != undefined ? activeElement?.data?.class : "";
+        htmlTag = activeElement?.data?.htmltag != undefined ? activeElement?.data?.htmltag : "span";
+
+        classInput.split(" ").forEach( cls => {
+
+            fontSizeList.forEach( fs => {
+                if(fs === cls.trim()) fontSize = fs;
+            });
+
+            textAlignmentList.forEach ( ta => {
+                if(ta === cls.trim()) textAlignment = textAlignmentList.indexOf(ta) ?? 0;
+            });
+
+            verticalAlignmentList.forEach( va => {
+                if(va === cls.trim()) verticalAlignment = va;
+            });
+            
+
+        });
+
+
+
     }
 
 
@@ -140,17 +180,221 @@
         if(activeElement){
             if(activeElement.data){
                 activeElement.data.class = classInput;
+                // classGlobal = classInput;
                 updateEditor();
             }
         }
     }
+
+
+    /**
+     * @typedef {Object} SelectOptions
+     * @property {string} value
+     * @property {string} name
+     * @property {string} info
+    */
+
+    /**
+     * Value of element
+     * @type Array<SelectOptions>
+     */
+    let fontSizeOptions = [
+        {name: "Default",   value: "",          info: "Page Default"},
+        {name: "X-Small",   value: "text-xs",   info: "font-size: 0.75rem; /* 12px */"},
+        {name: "Small",     value: "text-sm",   info: "font-size: 0.875rem; /* 14px */"},
+        {name: "Base",      value: "text-base", info: "font-size: 1rem; /* 16px */"},
+        {name: "Large",     value: "text-lg",   info: "font-size: 1.125rem; /* 18px */"},
+        {name: "X-Large",   value: "text-xl",   info: "font-size: 1.25rem; /* 20px */"},
+        {name: "2X-Large",  value: "text-2xl",  info: "font-size: 1.5rem; /* 24px */"},
+        {name: "3X-Large",  value: "text-3xl",  info: "font-size: 1.875rem; /* 30px */"},
+        {name: "4X-Large",  value: "text-4xl",  info: "font-size: 2.25rem; /* 36px */"},
+        {name: "5X-Large",  value: "text-5xl",  info: "font-size: 3rem; /* 48px */"},
+        {name: "6X-Large",  value: "text-6xl",  info: "font-size: 3.75rem; /* 60px */"},
+        {name: "7X-Large",  value: "text-7xl",  info: "font-size: 4.5rem; /* 72px */"},
+        {name: "8X-Large",  value: "text-8xl",  info: "font-size: 6rem; /* 96px */"},
+        {name: "9X-Large",  value: "text-9xl",  info: "font-size: 8rem; /* 128px */"},
+    ];
+
+    /**
+     * Actual font size class
+     * @type string
+     */
+    let fontSize;
+
+    /**
+     * List of all font size classes
+     * @type Array<string>
+     */
+    const fontSizeList = ["text-xs", "text-sm", "text-base", "text-lg", "text-xl", "text-2xl", "text-3xl", "text-4xl", "text-5xl", "text-6xl", "text-7xl", "text-8xl", "text-9xl",];
+
+    $: fontSize, (()=>{
+        if(loaded == true) {
+
+            var newClass = "";
+
+            classInput.split(" ").forEach( cls => {
+                switch(cls.trim()){
+                    case "text-xs":
+                    case "text-sm":
+                    case "text-base":
+                    case "text-lg":
+                    case "text-xl":
+                    case "text-2xl":
+                    case "text-3xl":
+                    case "text-4xl":
+                    case "text-5xl":
+                    case "text-6xl":
+                    case "text-7xl":
+                    case "text-8xl":
+                    case "text-9xl":
+                    // case " ":
+                        break;
+                    default:
+                        newClass += cls.trim();
+                        break;
+                }
+            });
+
+            classInput = newClass.trim() + " " + fontSize;
+            updateClass();
+        }
+    })();
+
+
+
+    /**
+     * Standard HTML tags that are applied to text element.
+     * @type Array<SelectOptions>
+     */
+    const htmlTags = [
+        {name: "h1 (Heading 1)",   value: "h1",          info: "Heading 1"},
+        {name: "h2 (Heading 2)",   value: "h2",          info: "Heading 2"},
+        {name: "h3 (Heading 3)",   value: "h3",          info: "Heading 3"},
+        {name: "h4 (Heading 4)",   value: "h4",          info: "Heading 4"},
+        {name: "h5 (Heading 5)",   value: "h5",          info: "Heading 5"},
+        {name: "h6 (Heading 6)",   value: "h6",          info: "Heading 6"},
+        {name: "div",              value: "div",         info: "div"},
+        {name: "span",             value: "span",        info: "span"},
+        {name: "p (Paragraph)",    value: "p",           info: "Paragraph"},
+    ];
+
+    /**
+     * Text of element
+     * @type string
+     */
+    let htmlTag;
+
+    $: htmlTag, (()=>{
+        if(loaded == true) {
+
+            if(activeElement){
+                if(activeElement.data){
+                    activeElement.data.htmltag = htmlTag;
+                    updateEditor();
+                }
+            }
+
+        }
+    })();
+
+
+    /**
+     * @type number
+     */
+    let textAlignment = 0;
+
+
+    /**
+     * List of all font size classes
+     * @type Array<string>
+     */
+    const textAlignmentList = ["text-left", "text-center", "text-right", "text-justify"];
+
+    $: textAlignment, (()=>{
+        if(loaded == true) {
+
+            var newClass = "";
+            classInput.split(" ").forEach( cls => {
+                switch(cls.trim()){
+                    case "text-left":
+                    case "text-center":
+                    case "text-right":
+                    case "text-justify":
+                        break;
+                    default:
+                        newClass += cls.trim();
+                        break;
+                }
+            });
+
+            classInput = newClass.trim() + " " + textAlignmentList[textAlignment];
+            updateClass();
+        }
+    })();
+
+
+    /**
+     * Vertical Alignment of element
+     * @type string
+     */
+    let verticalAlignment;
+
+    /**
+     * Vertical alignment options
+     * @type Array<SelectOptions>
+     */
+    const verticalAlignmentOptions = [
+        {name: "Default",       value: "",                  info: "Page Default"},
+        {name: "Baseline",      value: "align-baseline",    info: ""},
+        {name: "Top",           value: "align-top",         info: ""},
+        {name: "Middle",        value: "align-middle",      info: ""},
+        {name: "Bottom",        value: "align-bottom",      info: ""},
+        {name: "Text top",      value: "align-text-top",    info: ""},
+        {name: "Text bottom",   value: "align-text-bottom", info: ""},
+        {name: "Sub",           value: "align-sub",         info: ""},
+        {name: "Super",         value: "align-super",       info: ""},
+    ];
+
+    /**
+     * List of all font size classes
+     * @type Array<string>
+     */
+    const verticalAlignmentList = ["align-baseline", "align-top", "align-middle", "align-bottom", "align-text-top", "align-text-bottom", "align-sub", "align-super",];
+
+
+    $: verticalAlignment, (()=>{
+        if(loaded == true) {
+
+            var newClass = "";
+            classInput.split(" ").forEach( cls => {
+                switch(cls.trim()){
+                    case "align-baseline":
+                    case "align-top":
+                    case "align-middle":
+                    case "align-bottom":
+                    case "align-text-top":
+                    case "align-text-bottom":
+                    case "align-sub":
+                    case "align-super":
+                        break;
+                    default:
+                        newClass += cls.trim();
+                        break;
+                }
+            });
+
+            classInput = newClass.trim() + " " + verticalAlignment;
+            updateClass();
+        }
+    })();
+
 
 </script>
 
     <div class="widgetPanelSubTitle">Text Options</div>
     <br/>
 
-    <Optionsbutton items={["Content", "Design", "Animation"]} active={false} bind:value={selectedTabPageIndex}></Optionsbutton>
+    <Optionsbutton items={["Content", "Design", "Animation"]} bind:value={selectedTabPageIndex}></Optionsbutton>
 
     <div class="widgetPanelTabsDivider"></div>
 
@@ -161,21 +405,50 @@
 
     <div class="widgetPanelDivider"></div>
 
-    
-    <!-- <Textinput  bind:text={textInput} on:onSubmit={updateText} ></Textinput> -->
+    <div class="w-full flex flex-row flex-grow justify-between h-8 align-middle items-center">
+        <span title="Utilities for controlling the font size of an element.">Font Size</span>
+        <Select options={fontSizeOptions} bind:value={fontSize}/>
+    </div>
+
+    <div class="widgetPanelDivider"></div>
+
+    <div class="w-full flex flex-row flex-grow justify-between h-8 align-middle items-center">
+        <span title="Utilities for controlling the alignment of text.">Alignment</span>
+        <div class="w-[130px]">
+            <Optionsbutton items={['bi bi-text-left', 'bi bi-text-center', 'bi bi-text-right', 'bi bi-justify']} icons={true} bind:value={textAlignment}></Optionsbutton>
+        </div>
+    </div>
+
+    <div class="widgetPanelDivider"></div>
+
+    <div class="w-full flex flex-row flex-grow justify-between h-8 align-middle items-center">
+        <span title="Utilities for controlling the vertical alignment of an inline element.">V. Alignment</span>
+        <Select options={verticalAlignmentOptions} bind:value={verticalAlignment}/>
+    </div>
+
+    <div class="widgetPanelDivider"></div>
+
+    <div class="w-full flex flex-row flex-grow justify-between h-8 align-middle items-center">
+        <span title="Html container of text.">Html Tag</span>
+        <Select options={htmlTags} bind:value={htmlTag}/>
+    </div>
+
+    <div class="widgetPanelDivider"></div>
 
 
-    
+
  
-
-    
-
 
     {:else if selectedTabPageIndex==1}
 
     <div class="mb-1 w-full flex items-end align-bottom place-content-between"><span class="">Class</span><span class="text-[10px]">Tailwind CSS</span></div>
     <Textarea bind:text={classInput} on:onSubmit={updateClass} ></Textarea>
+    
+    <pre class="text-[8px]">{JSON.stringify(activeElement, null, 2)}</pre>
 
+    {:else if selectedTabPageIndex==2}
+
+    <pre class="text-[8px]">{JSON.stringify(activeElement, null, 2)}</pre>
    
     {/if}