+++ core/src/routes/menu/customize/fixed.svelte
@@ -1,5 +1,9 @@
 <script>
-
+   import { fade, fly } from 'svelte/transition';
+   /**
+    * @type string
+    */
+   export let side;
 </script>
 
 <div>