+++ core/src/routes/menu/options/fixed.svelte
@@ -1,5 +1,10 @@
 <script>
+   import { fade, fly } from 'svelte/transition';
 
+   /**
+    * @type string
+    */
+   export let side;
 
 </script>