+++ core/src/routes/menu/widgets/fixed.svelte
@@ -1,5 +1,11 @@
 <script>
+    import { fade, fly } from 'svelte/transition';
+    /**
+    * @type string
+    */
+    export let side;
 
+    // transition:fly="{{ x: side == "L" ? -320 : 320, duration: 300 }}"
 </script>
 
 <div>